Restructure of code. Created forward_models in ert folder.
