+++ setup.py
@@ -62,10 +62,10 @@ setup(
         "console_scripts": [
             "flownet=flownet._command_line:main",
             "flownet_render_realization=flownet.realization:render_realization",
-            "flownet_delete_simulation_output=flownet.ahm:delete_simulation_output",
-            "flownet_run_flow=flownet.ert._flow_job:run_flow",
-            "flownet_save_iteration_parameters=flownet.ahm:save_iteration_parameters",
-            "flownet_save_iteration_analytics=flownet.ahm:save_iteration_analytics",
+            "flownet_delete_simulation_output=flownet.ert.forward_models:delete_simulation_output",
+            "flownet_run_flow=flownet.ert.forward_models:run_flow",
+            "flownet_save_iteration_parameters=flownet.ert.forward_models:save_iteration_parameters",
+            "flownet_save_iteration_analytics=flownet.ert.forward_models:save_iteration_analytics",
             "flownet_plot_results=flownet.utils.plot_results:main",
         ],
     },
