+++ src/flownet/ahm/__init__.py
@@ -1,7 +1,2 @@
-from ._assisted_history_matching import (
-    AssistedHistoryMatching,
-    delete_simulation_output,
-    save_iteration_parameters,
-)
-from ._ahm_iteration_analytics import save_iteration_analytics
+from ._assisted_history_matching import AssistedHistoryMatching
 from ._run_ahm import run_flownet_history_matching
