+++ src/flownet/ahm/_assisted_history_matching.py
@@ -1,17 +1,8 @@
 import argparse
-import concurrent.futures
-import glob
-import json
-import os
-import pathlib
-import re
-import shutil
-from typing import List, Dict, Tuple
-
+from typing import List
 from configsuite import ConfigSuite
 import jinja2
 import numpy as np
-import pandas as pd
 
 from ..ert import create_ert_setup, run_ert_subprocess
 from ..realization import Schedule
@@ -160,97 +151,3 @@ class AssistedHistoryMatching:
                     else "      None      ",
                 )
         print("")
-
-
-def delete_simulation_output():
-    """
-    This function is called by a forward model in ERT, deleting unnecessary
-    simulation output files.
-
-    Returns:
-        Nothing
-
-    """
-    parser = argparse.ArgumentParser(prog="Delete simulation output.")
-
-    parser.add_argument(
-        "ecl_base", type=str, help="Base name of the simulation DATA file"
-    )
-
-    args = parser.parse_args()
-
-    for suffix in ["EGRID", "INIT", "UNRST", "LOG", "PRT"]:
-        if os.path.exists(f"{args.ecl_base}.{suffix}"):
-            os.remove(f"{args.ecl_base}.{suffix}")
-
-
-def _load_parameters(runpath: str) -> Tuple[int, Dict]:
-    """
-    Internal helper function to load parameter.json files in
-    parallel.
-
-    Args:
-        runpath: Path to where the realization is run.
-
-    Returns:
-        Dictionary with the realizations' parameters.
-
-    """
-    realization = int(re.findall(r"[0-9]+", runpath)[-2])
-    parameters = json.loads((pathlib.Path(runpath) / "parameters.json").read_text())
-
-    return realization, parameters["FLOWNET_PARAMETERS"]
-
-
-def save_iteration_parameters():
-    """
-    This function is called as a pre-simulation workflow in ERT, saving all
-    parameters of an iteration to a file.
-
-    The resulting dataframe is saved as a gzipped parquet file using a PyArrow table
-    and has the following format (example for 5 realizations and 2 parameters):
-
-    | index = realization | parameter 1 | parameter 2 |
-    |=====================|=============|=============|
-    |                   1 |         x.x |         x.x |
-    |                   3 |         x.x |         x.x |
-    |                   5 |         x.x |         x.x |
-    |                   4 |         x.x |         x.x |
-    |                   2 |         x.x |         x.x |
-
-    Mind that the dataframe is not ordered.
-
-    Returns:
-        Nothing
-
-    """
-    parser = argparse.ArgumentParser(prog="Save iteration parameters to a file.")
-    parser.add_argument("runpath", type=str, help="Path to the ERT runpath.")
-    args = parser.parse_args()
-    args.runpath = args.runpath.replace("%d", "*")
-
-    print("Saving ERT parameters to file...", end=" ")
-
-    iteration = int(re.findall(r"[0-9]+", sorted(glob.glob(args.runpath))[-1])[-1])
-    runpath_list = glob.glob(args.runpath[::-1].replace("*", str(iteration), 1)[::-1])
-    realizations_dict = {}
-
-    with concurrent.futures.ProcessPoolExecutor() as executor:
-        for result in executor.map(_load_parameters, runpath_list):
-            realizations_dict[result[0]] = result[1]
-
-    pd.DataFrame(
-        [parameters for _, parameters in realizations_dict.items()],
-        index=realizations_dict.keys(),
-    ).to_parquet(
-        f"parameters_iteration-{iteration}.parquet.gzip",
-        index=True,
-        engine="pyarrow",
-        compression="gzip",
-    )
-
-    shutil.copyfile(
-        f"parameters_iteration-{iteration}.parquet.gzip",
-        "parameters_iteration-latest.parquet.gzip",
-    )
-    print("[Done]")
