+++ tests/test_iteration_analytics.py
@@ -2,7 +2,7 @@ import math
 import numpy as np
 import pandas as pd
 
-from flownet.ahm._ahm_iteration_analytics import (
+from flownet.ert.forward_models._iteration_analytics import (
     prepare_opm_reference_data,
     prepare_flownet_data,
     normalize_data,
